Enable user comments
This feature is enabled modifing ALLOW_USER_COMMENTS in app_settings.py
and is disabled by default.
When is enabled, show a form to add comments from the users in
the detail of every book. The user do not need to be logued, then
a very simple captcha is added.