Add comments using django framework
This feature is enabled modifing ALLOW_USER_COMMENTS in app_settings.py
and is disabled by default.
When is enabled, show a form to add comments from the users in
the detail of every book. The user do not need to be logued.
It's implemented using the django comments framework [1]

[1] https://docs.djangoproject.com/en/1.4/ref/contrib/comments/